Move some code to utils and polish some functions
